Add files via upload
Review all the figures and tables that you generated in this assignment. 
Write at least three observations or inferences that can be made from the data. 
Include these observations at the top of your notebook.

- In general, the number of male and female mice tested are pretty equal
- In terms of Drug regiment, Capomulin is the most commonly used, the second one is Ramicane.
   The rest has the similar amount of drug regimen 
- Distribution of Tumor Volume (mm3) is pretty evenly

From calculation and graph
The lower quartile is: 48.72207785
The upper quartile is: 64.29983003
The interquartile is: 15.577752179999997
The the medianis: 59.85195552 
Values below 25.355449580000002 could be outliers.
Values above 87.66645829999999 could be outliers
-> no outliners